01-untrusted-data-boundary/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "poc-starter"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "anyio[trio]",
    "httpx",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "core*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
